Refactoriza controller
-Organizacion de codigo para trabajo colaborativo
